Increase maximum map size to 2^30 tiles (32k x 32k)
Raise MAX_MAP_TILES_BITS from 28 to 30. This required fixing every place
that assumed a tile index fits in <=28 bits or that packed a tile index
into a 32-bit value, plus the savegame paths that overflowed.

Save/load:
- map_sl.cpp: compute map chunk byte lengths in size_t, not uint32_t
  (size * 12 / size * 8 overflowed above 2^28, corrupting WMAP saves).
- Disable the combined WMAP chunk above the 4GB RIFF chunk limit, falling
  back to the per-field map chunks; widen the big-RIFF length read.
- Add XSLFI_LARGE_MAP_64BIT_KEYS (set only when Map::Size() > 2^28) to
  gate 64-bit key formats for the trace-restrict mapping (TRRM) and bridge
  signal style map (XBST) chunks, preserving the format for normal maps.

In-memory 64-bit tile packing:
- bridge signal style map, zoning sprite cache, trace-restrict ref id,
  YAPF rail segment cache key, AyStar (public roads / river gen) hash key.

Other:
- landscape.cpp: add 29-bit and 30-bit maximal-length LFSR feedback terms
  for the tile loop ordering.

Known limitation: the trace-restrict / programmable-signal GUI still packs
a tile index into a 32-bit WindowNumber (window identity only). 2^30 is the
ceiling until TileIndexDiff is widened from int32.

+++ src/bridge_signal_map.h
@@ -31,7 +31,7 @@ struct LongBridgeSignalStorage {
 
 extern robin_hood::unordered_flat_map<TileIndex, LongBridgeSignalStorage> _long_bridge_signal_sim_map;
 
-extern btree::btree_set<uint32_t> _bridge_signal_style_map;
+extern btree::btree_set<uint64_t> _bridge_signal_style_map;
 
 SignalState GetBridgeEntranceSimulatedSignalStateExtended(TileIndex t, uint16_t signal);
